+++ src/rcciconv.h
@@ -1,6 +1,6 @@
 #ifndef _RCC_ICONV_H
 #define _RCC_ICONV_H
 
-int rccIConv(rcc_context ctx, iconv_t icnv, const char *buf, int len);
+size_t rccIConv(rcc_context ctx, iconv_t icnv, const char *buf, size_t len);
 
 #endif /* _RCC_ICONV_H */
